Move test_did_doc_from_invitation_works
