+++ aries_vcx/src/did_doc/mod.rs
@@ -499,15 +499,4 @@ pub mod unit_tests {
         assert_eq!(String::from("1"), DidDoc::parse_key_reference(&_key_reference_1()));
         assert_eq!(_key_1(), DidDoc::parse_key_reference(&_key_1()));
     }
-
-    #[test]
-    fn test_did_doc_from_invitation_works() {
-        let mut did_doc = DidDoc::default();
-        did_doc.set_id(MessageId::id().0);
-        did_doc.set_service_endpoint(_service_endpoint());
-        did_doc.set_recipient_keys(_recipient_keys());
-        did_doc.set_routing_keys(_routing_keys());
-
-        assert_eq!(did_doc, DidDoc::from(Invitation::Pairwise(_pairwise_invitation())));
-    }
 }
